pyproject.toml
[tool.poetry]
name = "centrallockingsystem"
version = "0.1.0"
description = ""
authors = ["MennaFahmy"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2.2"
django-debug-toolbar = "^4.1.0"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
